perf(scattermoe): NVFP4+LoRA fp8-read split path (sm_120 memory win)
Follow-up to the NVFP4->bf16 fast path: keep the dequantized frozen expert
weight in fp8 (e4m3) instead of bf16 and read it in the grouped GEMM, upcasting
to bf16 in-register (bf16 MMA, bf16 activations). The fp8 transient is half the
bf16 one, so this halves the per-layer weight materialization: ~1.36x lower peak
memory (~6.3GB/layer at E=256) on the bandwidth-bound workstation expert GEMM,
and ~1.07x faster fwd+bwd -- a memory optimization that is roughly speed-neutral.

Key finding: fp8-read must use the SPLIT/non-fused path, not the fused LoRA
kernel. The fused kernel's two dots + extra LoRA A-tile load per K-iteration
serialize the in-register fp8->bf16 upcast and expose its latency (0.73x, a
regression). The base scatter2scatter (same routing, no LoRA) reads fp8 at
1.70x -- so it is the LoRA fusion, not the scatter structure, that breaks fp8.
fp8-read therefore routes through base-fp8 + bf16 LoRA (use_fused_bwd=False),
the opposite of the bf16 path.

- ops.py: base _compute_expert_block upcasts the weight tile to the activation
  dtype in-register (no-op for bf16; the half-bandwidth upcast for fp8).
- lora_ops.py: scatter2scatter_lora routes an fp8 weight to the split path.
- parallel_linear_lora.py: ScatterMoELoRA.forward skips the host bf16 upcast for
  an fp8 weight (the non-fused dX already reads the base weight via the now
  fp8-capable scatter2scatter).
- experts.py: _NVFP4_LORA_FP8_READ branch (env AXOLOTL_NVFP4_LORA_FP8_READ=1,
  default off) -> dequant NVFP4->fp8, use_fused_bwd=False, fp8 .recipe.
- selective_dequant_kernel.py: dequant_nvfp4_fp8_triton (one-pass NVFP4->fp8).

sm_120 only: datacenter (B200/B300) is compute-bound, fp8 ~= bf16, and the split
is slower than fused for bf16 -- so fp8-read would regress there. An arch gate
(_fp8_read_arch_ok) enforces this: off sm_120 the flag auto-falls-back to the
bf16-read path, so it is correct on both targets. NVFP4 only, default off. fp8
re-quant adds ~2.5% to the (already accepted)
NVFP4 weight error; no per-expert scale (it does not improve output quality).
Correctness at the fp8 floor (forward 3.2e-2, grads 3-9e-2 vs bf16-read); FSDP2
sharded grads bit-exact vs unsharded; new GPU-gated test for the fp8 dequant.

+++ src/axolotl/integrations/kernels/libs/scattermoe_lora/kernels/ops.py
@@ -55,7 +55,10 @@ def _compute_expert_block(
 
         X_blk_ptrs += BLOCK_K * stride_xk
         W_blk_ptrs += BLOCK_K * stride_wk
-        acc = tl.dot(x, w, acc, allow_tf32=allow_tf32)
+        # Upcast the weight tile to the activation dtype in-register. A no-op when W is already
+        # bf16/fp16; for an fp8 (e4m3) fp8-read weight it is the half-bandwidth upcast that lets
+        # the base grouped GEMM read fp8 and run a bf16 MMA (NVFP4+LoRA fp8-read fast path).
+        acc = tl.dot(x, w.to(x.dtype), acc, allow_tf32=allow_tf32)
     return acc
